requirements.txt
streamlit
pandas
pyyaml
requests
schedule
matplotlib